Add files via upload
This Java program serves as a student grade calculator, allowing users to input marks for six subjects: ARTS, MATHS, ENGLISH, CHEMISTRY, PHYSICS, and BIOLOGY. It employs a HashMap to store and manage the subject-wise marks. The user is prompted to provide marks on a scale of 0 to 100 for each subject. The program then calculates the total marks by summing up the individual subject marks and computes the average percentage. Based on the average percentage, the program assigns a grade, such as "S Grade," "A Grade," "B Grade," and so forth. The main method orchestrates these functionalities, displaying the total marks, average percentage, and the corresponding grade to the user. In essence, this code streamlines the process of evaluating and categorizing student performance, providing a quick overview of academic achievement across multiple subjects.